+++ Topic/Votes.go
@@ -2,14 +2,14 @@ package Topic
 
 // The Votes structure
 type Votes struct {
-	UpVotes   int `json:"upVotes"`
-	DownVotes int `json:"downVotes"`
-	SumVotes  int `json:"sumVotes"`
+	UpVotes   uint64 `json:"upVotes"`
+	DownVotes uint64 `json:"downVotes"`
+	SumVotes  int64  `json:"sumVotes"`
 }
 
 // GetSum is for calculate the Sum to SumVotes value
-func (v *Votes) GetSum() int {
-	return v.UpVotes - v.DownVotes
+func (v *Votes) GetSum() int64 {
+	return int64(v.UpVotes - v.DownVotes)
 }
 
 // SetSumVote is for set the sum value from GetSum
